Server support for All-Fields reporting.
Updated the existing AssessmentHtmlReport, AKA the SIS 1
All Fields report to work with SIS 2. Should work the exact
same as SIS 1 with a few minor improvements:

* Credited user listing should work properly
* Region listing should work properly
* Threats Classification Scheme should appear.
* Empty Regional Expert Questions should appear blank
* Standardized blank text to "(Not specified)", can be quickly changed
* Empty tables no longer drawn for data with no values.

Like the original incarnation, the report is driven based on
the existing SIS views -- so changes there will be reflected
in the reports, including re-organizing fields as well as section
headings.  Unlike its predecessor, though, the friendly display of
the data is no longer driven by the client-based structures, as they
are tightly-coupled in many cases with GWT functionality.  Instead,
this functionality is duplicated, re-using utility translation classes
where available.  Going forward, would be nice to share more of this
code, but for things like lookup values, having separate, optimized
solutions on both the client & server individually is a good idea.

Still, most of the code that is useful on both sides has now been
moved to clientshared.bridge (dealing with processing displays and
parsing XML).  Similar code moving has been done in other places to
allow for re-use; for example the AssessmentSchemaIO class was
created to allow for grabbing field XML without hitting the resource
directly.

Also, fixed a bug with the ViewParser that was skipping fields when
they were not declared in composite tags.  This was the cause of the
missing Threats fields on reports, and other fields could have been
skipped as well.

Additionally, fixed a couple issues in the existing "Blue" SIS 2 report:

* Changed table ID declarations to classes, as they are re-used
* Changed keywords to be CanonicalNames of fields for clarity in AssessmentReport.html
* Created subclass ReportTemplate to house useful functions for extensibility
* Simplified code calls in AssessmentReportTemplate

Finally, added a test case that runs the generators for these assessments
and saves a report on-disk for verification.  This will be fairly
difficult to create assertions for, but it could be done in the future.
There are currently three test cases:

* All Fields report based on the AllFields ("FullListing") view
* All Fields report based on NO view, just display fields with data
* SIS 2 "Blue" report

Ready to Deploy #701

Change-Id: I48800863f1be46145f1830832d55e6023a83b5a8
Reviewed-on: http://review.source.gogoego.com/3565
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>