LocalFileResource grabs assets for report HTML.
Added new LocalFileResource that can grab assets needed for
rendering the reports, particularly logo images and CSS files.
Changed the reports to look here instead of the old custom
restlets.

Also, changed the AssessmentReport to use the red list logo
instead of the IUCN logo on the "Blue" report.

Ready to Deploy #704

Change-Id: I4055b2440eb6533e9099e6d2d4fc14579acd06f0
Reviewed-on: http://review.source.gogoego.com/3567
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>