Ignore birds on taxon import.
Birds for taxa and assessments can now be ignored
when running import scripts.  This can be turned off
via the parameter bird=true, otherwise, the birds
will not be parsed.

Exclusion is based on the table "aves", which is now
in the scratch database, generated from the imported
database.  This table has the taxa ids of all birds,
determined by finding taxa within the class 'AVES'.

Ready to Deploy #571

Change-Id: Ib631d0ae37599848fa6de077a1cc45c1f6b319fe
Reviewed-on: http://review.source.gogoego.com/3571
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>