Constructor: HibernateLockRepository, FileLocker
Added a new constructor to HibernateLockRepository that takes
a SISPersistentManager object.  In practice, this can be used
for offline but it is not a long-term fix, as it breaks the
interface logic behind SIS.get().getLocker() ... which, in itself,
could use some change...

To support this at the API level, also added new constructor to
FileLocker.  In tandem, these can be used similarly to the IO
classes, placing API functionality atop a data target.

Usage:

   //Create objects
   HibernateLockRepository repository = new HibernateLockRepository(targetManager);
   FileLocker locker = new FileLocker(repository);

   //do lock things...
   locker.persistentEagerRelease(...);

See #645

Change-Id: I4ccacd5e05f93a97263cd4297334153ae744efe9
Reviewed-on: http://review.source.gogoego.com/3578
Tested-by: Rasanka Jayawardana <Rasanka.Jayawardana@iucn.org>
Reviewed-by: Rasanka Jayawardana <Rasanka.Jayawardana@iucn.org>