Layout corrections for Taxon General Info Page.
Added some margins to the content on the general information page
as well as moved the Edit button for TaxonomicNotes to the header
of the content panel to ensure the information in the pane scrolls
properly.

Also, set the spacing to the 50/50 instead of always showing all
information in the overview section, and added vertical scrolling
properties to both of these panels.  The overview section is also
collapsible if more room is needed to view the taxonomic notes.

Finally, re-sized the editing popup for taxonomic notes so that
the entire rich text area fits, and also scrolls when necessary.

Ready to Deploy #711

Change-Id: Id51cdaf71d3c83543b21b78846adf773821bf43a
Reviewed-on: http://review.source.gogoego.com/3584
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>