PermissionGroup authorization fix, remove old code
When checking the inheritance chain for permissions, the logic
will now apply a allow-based merging, so that if any permission
group has true permissions for a given operation, it will allow
that operation.  Previously, only the first instance of a given
permission would be taken into account.  With this method, one
will never encounter permission restrictions, only additions,
when applicable.

Fixed #714

Change-Id: Ibd15e9987667f32e1d06a4e4d5735f0c4a655917
Reviewed-on: http://review.source.gogoego.com/3595
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>