Hide legacy coding options for Class.Schemes.
Changed parsing for classification schemes to hide the
legacy coding options when using regular add.  This is
already done via quick add.  This was being attempted,
but was using old code known not to work.

Also with this change, optimized the parsing by checking
the root-level nodes to see if they are legacy, and if
so, eliding them completely from parsing.

Ready to Deploy #757

Change-Id: I9b8ba9997908acfccb8499dd6149a5cd98571d8e
Reviewed-on: http://review.source.gogoego.com/3648
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>