Schema-based, truly dynamic full report.
Made a few small adjustments to the AssessmentHtmlReport,
which drives the "full" and "available" versions of reports,
to allow for this to now be truly dynamic, as well as customizable
based on the assessment schema.

All views within a schema are required to have a view with the ID
of "Report".  This view will contain any fields that should be
in a report, and, of course, exclude the ones that should not appear.
The sections are also important here, re-ordering or re-naming a
section will cause that same ordering or naming change to appear in
the report.

From there, the reporting processor crawls the Reports view and
generates the report as normal, but it also inserts the proper logo
based on what the schema is (defaulting to the RedList logo).  In
this commit, a BirdLife logo is supplied for assessment created in
that schema.

Also, did a bit of code cleanup and refactoring.

See #763

Change-Id: Id586ebef4cbfbd40c028f4068e3fc427bc8ae1ea
Reviewed-on: http://review.source.gogoego.com/3650
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>