Reporting support for save to MS Word.
Added a version parameter which allows specifying that
a report should be saved as Word.  With this commit, the
logo will not persist, but the styling will.

Also, installed the Tribune font. While not in-use, it
can be easily applied by uncommenting a line in the CSS,
if this font is desirable.

Finally, when adding this new paramter to the UI, changed
the default for the limited set option to be true, and the
default for show empty fields to be false.

See #763

Change-Id: I8e64a7184dffc67fa7fc1d36957e667dc419e0eb
Reviewed-on: http://review.source.gogoego.com/3651
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>