Updated Access template, use HXTT, cleanup.
Updated the Access template file to use a 2002 version of the
software instead of an older version (probably 97).  Also, use
the HXTT drivers instead of JdbcOdbc, as those are not provided
across all Java installations.  Finally, added some changes to
the DBSession driver for HXTTAccessSession specifically in order
to deal with Boolean types.

Included in this commit is some small code cleanup and a test case
to run the access export standalone.

Ready to Deploy #778

Change-Id: I698030b67fd3fcfe6602cac699b8e2d3e592f3ca
Reviewed-on: http://review.source.gogoego.com/3688
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>