Access view builder to Viewmaster, improvements
A bit of code cleanup to the Viewmaster script to make it faster
and easier to run -- in a "set it and forget it" approach. Also,
added the new Access view builder to the process.  Finally, added
code to close the database connection.  This is important for
H2 databases.

See #778

Change-Id: Id64d6b62d2b7646947a10536e1de2dd94a96e564
Reviewed-on: http://review.source.gogoego.com/3689
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>