Prevent Infrarank setting for non-infrarank types.
Added code that eliminates the possibility of the infrarank
being set for species or higher, relying on the taxon level
instead of presence of an infrarank.  Any taxa that do have
this should be naturally cleansed with regular use.

Also, updated the conversion process to ensure that the
Infrarank never gets set for taxa being brought in.

Ready to Deploy #775

Change-Id: Id4570d56f7cfd46861a0ebed37b7d4107154a299
Reviewed-on: http://review.source.gogoego.com/3690
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>