Reports: Correct RedListCriteria parsing.
This field was being parsed as having no data because the structure
itself has no data.  Changed the parsing to not escape early for
this field.

Also, fixed the headings so that they are hard-coded to the display
strings from the UI, since they won't come over in the pretty data.

Ready to Deploy #783

Change-Id: Id62f820cf2ade35e9f590d7671618b4269cbbfd6
Reviewed-on: http://review.source.gogoego.com/3697
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>