GoGoEgo customizations to support SIS Offline
Added custom icons for Windows and Mac for SIS Offline,
and also updated the favicon to use the same icon; this
is the cropped animal from the red SIS logo.

Also, updated the com.solertium.util.restlet library to
support static and dynamic icon sets, such that an icon
provider can pass appropriate icons for the run context
via Files.  This allows SIS offline to package its icon
set with the software and have it display for users.
This has been extended to support both the TrayUI and
WindowUI implementations.

Finally, updated the equinox Application to look for a
new property, "GOGOEGO_DESKTOP", which will allow the
desktop mode to run without it having to be executed
via a Main method.

The properties to run this locally are being kept in the
instance.ini file alongside the product file.  While this
commit overrides GoGoEgo's properties, the changes made
to this project should be able to be pushed to a separate
project without much work.

See #29

Change-Id: Ib8fe6dc3abca72e694caff06ac5db01d1965e181
Reviewed-on: http://review.source.gogoego.com/3711
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>