OnLoad, ensure appropriate offline data available.
When initializing and loading data (offline metadata and
working set), to the client, ensure that the metadata and
the working set are properly loaded.  If either are not
loaded (this can happen when a user without right to the
offline working set attempts to login), prompt the user
with a message stating the error.

See #790

Change-Id: I37849b87f0e1817ece9bdf574ce38ea9c2518e24
Reviewed-on: http://review.source.gogoego.com/3718
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>