Display improvements for batch create assessment.
When creating assessments for a working set via the batch operation
(Create Draft Assessments), apply alphabetical sorting for taxa list
when choosing the "Selected Taxa" option.  Also, instead of displaying
this on-screen, use a popup window to provide more space, scrolling of
longer lists, and speed and layout improvements.

Ready to Deploy #769

Change-Id: Ied2679d08464b5dcbf3714ad6d36d9fd94f1d31e
Reviewed-on: http://review.source.gogoego.com/3725
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>