Hierarchical support for Not Assigned Taxa. (!)
The skinny: Nodes may now be entered into the taxon hierarchy as
"Not Assigned" by setting the Taxon Status accordingly. Taxa with
this status will not trigger as duplicates within Kingdoms, so
multiples of these can be inserted as needed.  When listed, these
taxon names will appear as light grey.

Included in this commit are optimizations to how the hierarchy is
downloaded to the client; old code had custom parsing for each
location needing to interrogate the hierarchy. This commit merges
this work into a single portable class, which remove excess code
and allows for future improvements for both speed as well as the
impending ticket for sequencing.

Because multiple taxa on a single level can have the same name,
search by name in level may no longer work in all cases.  So,
updated this functionality on the client to always search by ID
instead of the name.  This was happening via the "View Hierarchy"
button on the Taxon Home Page.

Finally, updated saving logic to ensure that when updating a taxa
from being Not Assigned to something else (say, New), the name
will be checked to ensure it's not breaking the rest of the
hiearchy.

A single insert is needed to add support for this feature:

   INSERT INTO taxon_status(id, "name", code) VALUES (7, 'Not Assigned', 'I');

The 'I', here, is for 'Incertae sedis', or uncertain placement.

Ready to Deploy #756

Change-Id: Ie405a10960b6fac3233d692ed899e7d1b8dc1352
Reviewed-on: http://review.source.gogoego.com/3727
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>