Code cleanup and fixes for offline importer.
Renamed instances of "source" and "target" sessions to
"offline" and "online", respectively, for code clarity.

Also, assessment changes were not being committed, hence
the Edit errors.  Start and commit a new transaction for
each individual assessment sync. Additionally, print
additional information when debugging.

Finally, changed the Manifest to 2.0 and name to Offline
Support, and added restful means of specifying that the
locks should NOT be released upon import.

See #645

Change-Id: I919d8a8be35fd5462a5c6a6a6e6a15d29c8dceb1
Reviewed-on: http://review.source.gogoego.com/3738
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>