Reduce document parsing for AssessmentHtmlTemplate
Took a few shortcuts in order to make for less parsing and
serializing of the view document.  The old way was performing
one extra parse and serialization when building the document.
This should help with ensuring there aren't IO errors when
reading the view document for reports.

This is necessary due to the forthcoming changes to the report;
any cached documents on the VFS for SIS instances need to be
removed.

See #780

Change-Id: If57d936a7b10a4d7ffa7a9cece271ea31678fc77
Reviewed-on: http://review.source.gogoego.com/3745
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>