SIS Global Settings now drive SIS Application
Moved the SIS-specific settings from instance.ini to the
SIS settings file global.properties.  This will make settings
management easier and dynamic since it can be managed via the
GoGoEgo Application Manager.  This also means that they can
be interacted with from with-in the application, which is
important for Offline.

Also, applied code cleanup to WorkingSetExporter.

Fixed #807

Change-Id: Ia7ee324c8e0f294a36f23ab01c84e5c5bad62975
Reviewed-on: http://review.source.gogoego.com/3747
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>