Correct save and merge functions for WSFilter.
For assessment types (drafts, published, etc), the working set
filter panel was not properly setting the published assessment
type.  Therefore, some working sets were being saved with the
published assessment type selected.  Updated this to not even
allow this to happen unless the button appears on-screen.

Also, fixed the merge-parser for WorkingSet to ensure that the
assessment types get properly merged.

See #800

Change-Id: I3f6cf580ae054c6161baa9244e280e6c0a4e161e
Reviewed-on: http://review.source.gogoego.com/3761
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>