Persist Cache-Control headers through GWTPProxy.
The GWT-Aware Persistent Proxy removes all headers from
response entities, but this causes issues with browsers
that aggressively cache, such as the new Firefox.  To
combat this, the proxy will allow for the Cache-Control
header to remain in the headers list if it was specified
in the original response.

Also, changed the logging properties for Restlet to level
SEVERE, which should keep the noise level to a minimum in
the GWT console.

Change-Id: Iea51bdfe562bc2e494949dfd393008f4e6aa3512
Reviewed-on: http://review.source.gogoego.com/3765
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>