Synchronize software update from source to Offline
Offline SIS can check for software updates via the manager and
merge them in.  This can also be run in test mode, which will
be of use later.

Also, data synchronization now depends on your software being up
to date; if updates are found, you will not be allowed to sync
until you update.

Ready to Deploy #819

Change-Id: I27a882f073a0da44a95d88020967ddfce5ed7065
Reviewed-on: http://review.source.gogoego.com/3795
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>