Updated reference search query to return usage count.
The reference search query will now return the usage counts for
all known data types, including:

* Common Name
* Synonym
* Taxon
* Field
* Assessment

It will also return the sum of the individual usages as a total
in the total_count attribute.

Additionally, the SIS 1 behavior has been re-instituted, showing
the system-wide usage of the references as opposed to the context-
sensitive count. The performance hit is negligible for the online
PostgreSQL DB, but noticeable in offline H2.  The context-sensitive
behavior can be easily reinstituted following a TODO in the code.

While this does not solve ticket #747 or problems with the reference
client display, it should ease some immediate pains while we work
to devise a better overall solution.

Ready to Deploy #830

Change-Id: Iabd2901045cb0b43db87a524187fb4ff996ab9fb
Reviewed-on: http://review.source.gogoego.com/3806
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>