Foreign data mode for AssessmentPersistence.
Implemented a new mode for AssessmentPersistence to deal with
handling foreign data (in this particular case, data coming
from Offline).

If a field is found with an ID but it does not exist in that
assessment, rather than discard it, allow a deep copy of that
field to be added.

Test #833

Change-Id: Iaff11dd1435aade9e55624a332790ef3a0604db3
Reviewed-on: http://review.source.gogoego.com/3807
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>