Bulk User Import: swap mid_initial for initials.
Importer no longer takes middle initial in the CSV, instead
it takes simply the user's initials.  Initials are also no
longer generated -- if the initials should be something other
than what SIS already generates for citations, it will need
to entered here.

Also, added a test case to make sure this was working for
the initials portion as advertised.

Ready to Deploy #824

Change-Id: Ib1c96962be827eccb2ac92df0625b3bad607253c
Reviewed-on: http://review.source.gogoego.com/3811
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>