StressesConverter, FieldMigrationConverters.
Refactored the latest batch of converters used for migration
fixes into FieldMigrationConverters.  All follow a pattern of
first generarting a table with the corrections it would make
for analytical purposes.  Once all checks out OK, a second
mode allows the updates to actually process based on this
generated table.

Also, added a converter for Stresses.  Though they converted
correctly the first time around, the data model linkage was
not properly done, causing Threats to appear with no stresses.

Finally, fixed the spot in the AssessmentConverter where the
data linkage was not being made.

Ready to Deploy #827

Change-Id: I55936a94096f27e9d0e6d7c8b0b8ae8f3152ba8b
Reviewed-on: http://review.source.gogoego.com/3812
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>