Support for tabular view for OneToMany fields.
Added support for a tabular display of the data within
one-to-many fields.  They will display similarly to how
classification schemes appear -- with the data in view
format, without having to click the View or View/Edit
button first.

To use this, need to update the appropriate field XML
and add a "columns" attribute to the oneToMany node.
Here, declare the numeric indexes of which fields should
be displayed in the table in comma-separated format.

See #773

Change-Id: I62aed9d4dc816e0fb5b0b2ae0019a94680269fd8
Reviewed-on: http://review.source.gogoego.com/3814
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>