Restored support for offline update detection.
Updates can now be tested on the client login page as
in the 1.0 version of Offline.  Updating will not
happen automatically, though; will still need to go to
the offline manager for that step.

In doing this, cleaned up the code a bit, removed old
code, and added new property to SIS index.html to take
the software version.

Ready to Deploy #835

Change-Id: Ia5b0e13f8f1018bbe603c27798abc4800aef3a38
Reviewed-on: http://review.source.gogoego.com/3817
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>