Offline support for ImageViewerRestlet.
While images are still not supported in the online context,
the image viewer restlet has been added as an offline target,
returning the not-available image and preventing the HTTP
BASIC login popup.

Also, added a warning in the client stating that this function
is not available offline.  Would be nice going forward to have
an image saying this instead, but this will work in the interim.

Fixed #831
Ready to Deploy #836

Change-Id: I75c65b598ccde97c474cca05e6be5ff2dcd46745
Reviewed-on: http://review.source.gogoego.com/3824
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>