Added isScoped method to AuthorizableObject API.
Objects can now determine whether or not they should fall
under scoping rules.  This information was previously
determined at the permission checking level, which makes it
difficult to add new items to the API.  This change will make
it more clear when implementing new authorizable objects.

Of course, in doing this, updated all existing instances of
AuthorizableObject with the correct implementation for this
new method.

Ready to Deploy #845

Change-Id: I8a7f83a21fd18f75edd606e190f0f9f8252d7281
Reviewed-on: http://review.source.gogoego.com/3888
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>