Delete data from hidden fields on save.
When saving fields, if the field has been hidden due to
being in a related structure, that data should be removed
from the database.

Also, classification schemes should update their views on
save to ensure it stays up-to-date.  With this commit, due
to the nature of the static tables, data won't be automatically
updated upon selection; it will, however, update once data
is saved to the database.

Ready to Deploy #853

Change-Id: Ic09f98d8ed3784174231517ae296c9341d30217b
Reviewed-on: http://review.source.gogoego.com/3900
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>