Installed Integrity Validation in Publication Workflow.
Integrity validator will now run when submitting assessments or working
sets for publication IF the appropriate setting is enabled.  The new
SIS setting is:

   org.iucn.sis.server.publication.validate

Setting this to true will trigger the validation.  Failure will stop
the submission process and prompt the user to view the full, detailed
results in a new window.

Ready to Deploy #688

Change-Id: Id9dd6ac446e9078d3319147150def17d4a6c05d4
Reviewed-on: http://review.source.gogoego.com/3909
Tested-by: Carl Scott <carl.scott@solertium.com>
Reviewed-by: Carl Scott <carl.scott@solertium.com>