OnLogout, remove listeners from UserStore cache.
Under certain circumstances, an infinite loop was possible when
logging and and logging in again while accessing the UserViewPanel.
This was due to the UserStore not removing the approprite store
listeners in its onLogout function.

Added the appropriate calls to remove these listeners to ensure
proper behavior for the next login.

Ready to Deploy #846

Change-Id: If20d4bc4d7378eed2388661a6f211f19baab6f90
Reviewed-on: http://review.source.gogoego.com/3953
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>