Create separate SIS equinox application.
Standalone SIS equinox application with adjustments made
such that it is not adversely affected if/when GoGoEgo
changes are merged in.

Also, corrected some Manifest issues with the Apache Commons
plugins that were preventing correct builds.  In addition,
the commons libraries, among others, are now standard plugins
in the SIS build, which helps with some "plugins" that rely
on reflection of these common classes.

Finally, since it was convenient, removed some annoying print
statements from GoGoEgo.

This is needed for the DEM build to make the legacy plugins,
with issues highlighted above, work in SIS 2.

See #591

Change-Id: I6e4da44f388d904fe647d3d0846da87ede9c6d1c
Reviewed-on: http://review.source.gogoego.com/3968
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>