Clean up batch create reporting XHTML.
The working set batch create assessmetn success reporting text
was pretty bad; an unsorted CSV of taxa names.  Changed this to
use valid XHTML with the taxon names in alphabetical order, as
well as listing the number of taxa in each reporting section (for
success, failure, or conflict).  Should help with larger working
sets.

Change-Id: I39f072de6f2d5c03cd1d1d04817f495bb1650db0
Reviewed-on: http://review.source.gogoego.com/4036
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>