Access Copier via Jackcess.
The access copier will create a new database schema for
each Access Export request.  The schema will create a vw_filter
table containing either (a) taxa in the given working set or
(b) assessments in the giving working set.  If (b) is chosen, then
only assessments fitting the working set scope will be available
in that export.  Option (a) is looking ahead to future requirements,
but other factors need to be evaluated before exposing this.

Once the new schema and vw_filter table gets created, the field-based
views will be created, as per usual -- except as tables.  This is
because these queries run fast anyway, and other views can be built
atop these tables.  Following that, views will be created for the
"simplified" queries, hopefully atop the newly-generated tables.
First utility and base queries, then the "public" queries. Note: any
queries prefixed with "vw_" will not be transferred in the export.

Once all the views and tables have been generated, the result will
be tables (and view) that contains *exactly* the data requested by
the user, so all that need to be done is a simple, literal, table
copy from this schema to the awaiting Access database.  No extra
queries needed.

TODO:

* Allow user choice of filtered vs unfiltered export.
* Correct query for StressesSubfield.

Ready to Deploy #891

Change-Id: I51eca2785ca4f9d9b8d063e602071294bffb1a99
Reviewed-on: http://review.source.gogoego.com/4115
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>