Historical assessment support for Access Export.
Added support for exporting information for historically published
assessments with the Access Export.  Also, implemented a dialog
window to allow for these settings to be exposed in the client.
This dialog is current hard-coded to contain the sole access format
(for SIS 2.0), but this can later be expanded easily to be completely
server-driven.

Also, updated the test cases to generate historical versions of
exported working sets.

Ready to Deploy #897

Change-Id: I63b642a8660dbe0e4c002c59d84cf6d01ee34d4b
Reviewed-on: http://review.source.gogoego.com/4192
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>