Viewmaster fixes for Credited Users, Stresses
Fixed an issue with the stresses table where the recordid
was not appearing correctly, and overall the data was not
being correctly populated.  Created separate table to build
this; in the future, the recursive building process should
take care of this.

Also, fixed issue with data concatenation for the credit
user fields, such that now all data will appear on one
row in a comma-separated list.

See #946

Change-Id: Ia6449d7b94c09c61c0dc511c0b3d740acda14c0d
Reviewed-on: http://review.source.gogoego.com/4267
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>