Viewmaster: additional scripts fixes and additions.
Added new scripts to the additional scripts to generate publication
views for both IAS taxa as well as viruses.  Also, changed the
generated lookups tables to be uppercase like the other lookup
tables.

Finally, the IAS lookup table pulls from the taxon table and correctly
prepends "Unspecified" to the name of taxa for level Genus and above.

See #946

Change-Id: I93c6823aa06de0aab2c353ee791cfbeae0a26a3f
Reviewed-on: http://review.source.gogoego.com/4288
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>