Quickie adjustments for BirdLife databases.
Added an adjustment script for all things BirdLife that need to happen
in addition to normal running context.  This script can be triggered by
including it by filename via the viewmaster.additional.include property.

This script will replace the threatssubfield_threatslookup view with a
view that correctly pulls information from the adjusted view table,
such that items marked as 8.1.1 when they should be 8.1.2 will appear
correctly.  This overwrites the correctly-generated views from the
additional builder, so once the data is correct, we can simply remove
the file entirely.

In addition, the publication view for IAS taxa is cleaned up to exclude
the common name, for consistency.

See #946

Change-Id: I0815682e0d21d321375f21b56aff5f2236c4f5ca
Reviewed-on: http://review.source.gogoego.com/4289
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>