Incorporate Unknown in A2, A4 calculations.
A2 and A4 criteria rules allow for unknown being factored into
the calculator.  Changed the API here a bit to allow for
specifying how to deal with unknowns, and for A2 and A4 to
treat unknown as "0-1".  By default, unknowns are treated as
null.

Also, added a test cases for this to ensure all is well.

Ready to Deploy #905

Change-Id: I371883b219d486de42e948ee92b20d40d82726cc
Reviewed-on: http://review.source.gogoego.com/4293
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>