Offline: Catch errors caused by missing username.
When no username is supplied for synchronizing data,
or if a username is supplied, but that name is not
in the database, throw an error alerting the user of
the issue, stopping the sync process.

Also, trim the username such that leading and trailing
whitespace (common when copying and pasting) will not
cause issues.

Ready to Deploy #967

Change-Id: I9e510ae619c50dc91aa4a56b8fe7fec195c8c5f5
Reviewed-on: http://review.source.gogoego.com/4345
Reviewed-by: Carl Scott <carl.scott@solertium.com>
Tested-by: Carl Scott <carl.scott@solertium.com>